requirements.txt
schedule==1.2.2
python-dotenv==1.1.1
psycopg2-binary==2.9.11
python-telegram-bot==13.7